Move common functionality to ConfigManager and define interface
